+++ src/test/java/mil/nga/geopackage/test/GeoPackageImportTest.java
@@ -109,14 +109,4 @@ public class GeoPackageImportTest extends ImportGeoPackageTestCase {
 
 	}
 
-	/**
-	 * Test draw feature preview
-	 */
-	@Test
-	public void testDrawFeaturePreview() {
-
-		GeoPackageTestUtils.testDrawFeaturePreview(geoPackage);
-
-	}
-
 }