+++ src/test/java/mil/nga/geopackage/test/features/FeaturePreviewImportTest.java
@@ -1,0 +1,36 @@
+package mil.nga.geopackage.test.features;
+
+import java.io.IOException;
+
+import org.junit.Test;
+
+import mil.nga.geopackage.test.ImportGeoPackageTestCase;
+
+/**
+ * Test Feature Preview from an imported database
+ * 
+ * @author osbornb
+ */
+public class FeaturePreviewImportTest extends ImportGeoPackageTestCase {
+
+	/**
+	 * Constructor
+	 */
+	public FeaturePreviewImportTest() {
+
+	}
+
+	/**
+	 * Test draw
+	 * 
+	 * @throws IOException
+	 *             upon error
+	 */
+	@Test
+	public void testDraw() throws IOException {
+
+		FeaturePreviewUtils.testDraw(geoPackage);
+
+	}
+
+}