+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Adheres to [Semantic Versioning](http://semver.org/).
 * XYZ tile rebranding, previously referred to as standard and/or Google
 * Separation of columns from user tables
 * Queries by specified columns
+* Feature Style Extension fix to use an individual feature style over a default table icon
+* GeoPackageImpl and FeatureIndexGeoPackageResults classes changed to be public
 
 ## [3.4.0](https://github.com/ngageoint/geopackage-java/releases/tag/3.4.0) (11-14-2019)