+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 ## 3.5.0 (TBD)
 
 * geopackage-core version 3.5.0
+* sqlite-jdbc version 3.30.1
 * Tile Generator option to skip pre-existing tiles when updating a GeoPackage
 * URL Tile Generator continuation by skipping existing tiles (option to force replace)
 * XYZ tile rebranding, previously referred to as standard and/or Google